+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod buffer;
 pub mod error;
 pub mod fs;
+pub mod hints;
 pub mod inode;
 pub mod sparse_cache;